# docs/specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LoonFS HTTP API",
    "description": "Static OpenAPI document for the LoonFS v0 HTTP API.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.2.1"
  },
  "paths": {
...
    "/v0/namespaces/{namespace_id}/inodes/{inode_id}": {
      "get": {
        "tags": [
          "inodes"
        ],
        "summary": "Stat inode",
        "description": "Returns the current path entry for a visible inode. Unknown or hidden inodes answer `inode_not_found`.",
        "operationId": "stat_inode",
        "parameters": [
          {
            "name": "namespace_id",
            "in": "path",
            "description": "Namespace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "inode_id",
            "in": "path",
            "description": "Inode ID",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^ino_[1-9][0-9]*$"
            },
            "example": "ino_123"
          },
          {
            "name": "include_attributes",
            "in": "query",
            "description": "Project the inode's attribute map and revision (`true` or `false`). Defaults to `true`: a stat answers for one path and a map is capped at 64 KiB.",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authoritative current inode entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthoritativePathEntry"
                }
              }
            }
          },
          "400": {
            "description": "Invalid inode ID or include_attributes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace or visible inode not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "408": {
            "$ref": "#/components/responses/DeadlineExceededResponse"
          },
          "410": {
            "description": "Namespace deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v0/namespaces/{namespace_id}/inodes/{inode_id}/revisions": {
      "get": {
        "tags": [
          "inodes"
        ],
        "summary": "List file revisions by inode",
        "description": "Returns retained revisions for a file inode without requiring a current path.",
        "operationId": "list_file_revisions_by_inode",
        "parameters": [
          {
            "name": "namespace_id",
            "in": "path",
            "description": "Namespace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "inode_id",
            "in": "path",
            "description": "File inode ID",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^ino_[1-9][0-9]*$"
            },
            "example": "ino_123"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum page size",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1000,
              "maximum": 1000,
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque file-revisions page cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File revisions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListFileRevisionsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid inode ID, limit, or cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace or inode not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "408": {
            "$ref": "#/components/responses/DeadlineExceededResponse"
          },
          "409": {
            "description": "Inode is not a file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "410": {
            "description": "Namespace deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v0/namespaces/{namespace_id}/inodes/{inode_id}/revisions/{revision_no}/content": {
      "get": {
        "tags": [
          "inodes"
        ],
        "summary": "Read file revision by inode",
        "description": "Reads and verifies one retained file revision by inode ID and revision number.",
        "operationId": "get_file_revision_bytes_by_inode",
        "parameters": [
          {
            "name": "namespace_id",
            "in": "path",
            "description": "Namespace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "inode_id",
            "in": "path",
            "description": "File inode ID",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^ino_[1-9][0-9]*$"
            },
            "example": "ino_123"
          },
          {
            "name": "revision_no",
            "in": "path",
            "description": "Revision number",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RevisionNo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revision bytes",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32",
                    "minimum": 0
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid inode ID or revision number",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace, inode, or revision not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Inode is not a file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "410": {
            "description": "Namespace deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Content exceeds the advertised `download.max_content_bytes` limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "The server is at its concurrent content-read limit; retry shortly",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AbsolutePath": {
        "type": "string",
        "description": "Validated complete absolute namespace path, serialized as a plain string.",
        "example": "/docs/report.txt"
      },
...
      "ActorRef": {
        "type": "object",
        "description": "Identifies the user, service, or system responsible for a commit.\n\nLoonFS stores this value as provided. It does not authenticate the actor or\nlook up profile information.",
        "required": [
          "kind",
          "id"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ActorId",
            "description": "A stable identifier supplied by the application."
          },
          "kind": {
            "$ref": "#/components/schemas/ActorKind",
            "description": "The type of actor."
          }
        },
        "additionalProperties": false
      },
...
      "ApiError": {
        "type": "object",
        "description": "HTTP error body used by LoonFS APIs.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable reason from the [`ErrorCode`](crate::ErrorCode)\nregistry.\n\nCarried as a string so clients keep working when a newer server\nintroduces a code they do not know; use\n[`ErrorCode::parse`](crate::ErrorCode::parse) for typed access."
          },
          "details": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ErrorDetails",
                "description": "Structured context for the code, present when the failure carries\nmachine-usable identity (API spec, \"Standard error contract\"). Boxed\nso the rare detailed error does not widen every error-carrying result."
              }
            ]
          },
          "feature": {
            "type": [
              "string",
              "null"
            ],
            "description": "For `not_supported` errors, the capability-document feature key the\nclient should reconcile against."
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message."
          },
          "param": {
            "type": [
              "string",
              "null"
            ],
            "description": "Identifies the invalid input. Body fields use JSON Pointer paths;\nquery and path parameters use their names; CLI errors use the flag or\nargument as written."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Correlation id the server assigned to the failed request; the same\nvalue is sent as the `x-request-id` response header."
          }
        }
      },
...
      "AttributesProjection": {
        "type": "object",
        "description": "One inode's structurally complete attribute projection.",
        "required": [
          "attributes_revision_no",
          "attributes"
        ],
        "properties": {
          "attributes": {
            "$ref": "#/components/schemas/Attributes",
            "description": "The complete attribute map at `attributes_revision_no`.\n\nAn inode that has never had attributes written is at revision 0 with\nan empty map."
          },
          "attributes_revision_no": {
            "$ref": "#/components/schemas/AttributeRevisionNo",
            "description": "The attribute revision this projection represents."
          },
          "attributes_updated_at_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Time of the latest attribute update, in Unix milliseconds. This is\n`None` for the initial empty state at revision 0.",
            "minimum": 0
          },
          "attributes_updated_by": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ActorRef",
                "description": "Actor responsible for the latest attribute update. This is `None` for\nthe initial empty state at revision 0."
              }
            ]
          }
        }
      },
      "AuthoritativePathEntry": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AuthoritativePathEntryKind",
            "description": "File-or-directory classification and its kind-specific payload."
          },
          {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AttributesProjection",
                "description": "The inode's attribute projection, when requested."
              }
            ]
          },
          {
            "type": "object",
            "required": [
              "namespace_id",
              "path",
              "inode_id",
              "created_by",
              "created_at_ms",
              "head_seq"
            ],
            "properties": {
              "created_at_ms": {
                "type": "integer",
                "format": "int64",
                "description": "Time the inode was created, in Unix milliseconds. Sequence numbers\ndetermine order.",
                "minimum": 0
              },
              "created_by": {
                "$ref": "#/components/schemas/ActorRef",
                "description": "Actor that created this inode, as supplied by the application."
              },
              "display_name": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/DisplayName",
                    "description": "Stored display name for this path component, absent for the nameless root."
                  }
                ]
              },
              "head_seq": {
                "$ref": "#/components/schemas/ChangeSeq",
                "description": "Namespace head sequence this answer was read from."
              },
              "inode_id": {
                "type": "string",
                "description": "Stable inode ID within a namespace",
                "example": "ino_123",
                "pattern": "^ino_[1-9][0-9]*$"
              },
              "namespace_id": {
                "$ref": "#/components/schemas/NamespaceId",
                "description": "Namespace that was read."
              },
              "parent_inode_id": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "type": "string",
                    "description": "Stable inode ID within a namespace",
                    "example": "ino_123",
                    "pattern": "^ino_[1-9][0-9]*$"
                  }
                ]
              },
              "path": {
                "$ref": "#/components/schemas/AbsolutePath",
                "description": "Absolute path as rendered from stored display names."
              }
            }
          }
        ],
        "description": "Authoritative metadata for one visible path.\n\nThis is the result shape for stat/list style reads. The entry kind carries\nthe file-only revision and content summary, so a directory cannot carry a\npartial file payload. Attributes are likewise projected as one value or\nomitted as one value, while serializing as prefixed sibling fields.\nThe attribute revision is read independently — clients feed it to\n`expected_attributes_revision_no` on the next write without touching the\nvalues — so this is a prefixed-sibling projection rather than a value\nconsumed as one nested unit."
      },
      "AuthoritativePathEntryKind": {
        "oneOf": [
          {
            "type": "object",
            "title": "AuthoritativePathEntryDirectory",
            "description": "A directory, which has no revision payload in v0.\n\nThe entry tag reuses [`InodeKind`]'s wire vocabulary.",
            "required": [
              "inode_kind"
            ],
            "properties": {
              "inode_kind": {
                "type": "string",
                "enum": [
                  "dir"
                ]
              }
            }
          },
          {
            "type": "object",
            "title": "AuthoritativePathEntryFile",
            "description": "A file and its current revision summary.",
            "required": [
              "revision_no",
              "size_bytes",
              "content_ref",
              "revision_committed_by",
              "revision_committed_at_ms",
              "inode_kind"
            ],
            "properties": {
              "content_ref": {
                "$ref": "#/components/schemas/ContentRef",
                "description": "Current content reference."
              },
              "inode_kind": {
                "type": "string",
                "enum": [
                  "file"
                ]
              },
              "revision_committed_at_ms": {
                "type": "integer",
                "format": "int64",
                "description": "Time of the current revision, in Unix milliseconds. Revision\nsequences determine order.",
                "minimum": 0
              },
              "revision_committed_by": {
                "$ref": "#/components/schemas/ActorRef",
                "description": "Actor responsible for the current revision."
              },
              "revision_no": {
                "$ref": "#/components/schemas/RevisionNo",
                "description": "Current file revision number."
              },
              "size_bytes": {
                "type": "integer",
                "format": "int64",
                "description": "Current file size in bytes.\n\nThis remains explicit even though `content_ref` also carries the\nlength because callers sort directory listings by this field.",
                "minimum": 0
              }
            }
          }
        ],
        "description": "Kind-specific metadata for an authoritative path entry."
      },
...
      "ChangeSeq": {
        "type": "integer",
        "format": "int64",
        "description": "Sequence number assigned to a namespace commit. It determines the order in which commits become visible.",
        "maximum": 9007199254740991,
        "minimum": 0
      },
...
      "ContentRef": {
        "type": "object",
        "description": "Pointer to one immutable content object.\n\n`content_id` is identity — *which* object — and the checksum is\nevidence about its bytes. Separating the two is what lets the final\nobject key exist before the first byte is read.\n\nA `ContentRef` is safe to publish only after the referenced bytes are\ndurable in the namespace's content store.",
        "required": [
          "kind",
          "content_id",
          "size_bytes",
          "checksum"
        ],
        "properties": {
          "checksum": {
            "$ref": "#/components/schemas/Checksum",
            "description": "Mandatory checksum over the complete object."
          },
          "content_id": {
            "$ref": "#/components/schemas/ContentId",
            "description": "Immutable identity of the referenced object."
          },
          "kind": {
            "type": "string",
            "description": "Content strategy used by the referenced object."
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64",
            "description": "Complete byte length of the referenced content.",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
...
      "DisplayName": {
        "type": "string",
        "description": "User-facing spelling of one path component.",
        "example": "report.txt"
      },
...
      "ListFileRevisionsResponse": {
        "type": "object",
        "description": "Response for listing file revisions.",
        "required": [
          "namespace_id",
          "inode_id",
          "head_seq",
          "revisions"
        ],
        "properties": {
          "head_seq": {
            "$ref": "#/components/schemas/ChangeSeq",
            "description": "Namespace head sequence used for the read."
          },
          "inode_id": {
            "type": "string",
            "description": "Stable inode ID within a namespace",
            "example": "ino_123",
            "pattern": "^ino_[1-9][0-9]*$"
          },
          "namespace_id": {
            "$ref": "#/components/schemas/NamespaceId",
            "description": "Namespace that was read."
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque cursor for the next page, if more revisions are available."
          },
          "revisions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileRevision"
            },
            "description": "Retained revisions in order."
          }
        }
      },
...
      "NamespaceId": {
        "type": "string",
        "description": "Durable id for one namespace.\n\nA namespace is one filesystem history. This id is not a display name and\nshould not be reused after destruction. Its serialized form is 1 to 128\nlowercase ASCII letters, digits, dots, underscores, or hyphens, starting\nwith a letter or digit; the `loonfs-` prefix is reserved for system use.",
        "example": "demo",
        "pattern": "^[a-z0-9][a-z0-9._-]{0,127}$"
      },
...
      "RevisionNo": {
        "type": "integer",
        "format": "int64",
        "description": "Revision number for a file's content. It increases whenever the content is replaced or restored.",
        "maximum": 9007199254740991,
        "minimum": 0
      },
...
    },
    "responses": {
      "DeadlineExceededResponse": {
        "description": "The request exceeded `request_deadline_ms`. A timed-out mutation may still complete, so clients must determine its outcome before retrying.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "description": "HTTP error body used by LoonFS APIs.",
              "required": [
                "code",
                "message"
              ],
              "properties": {
                "code": {
                  "type": "string",
                  "description": "Stable machine-readable reason from the [`ErrorCode`](crate::ErrorCode)\nregistry.\n\nCarried as a string so clients keep working when a newer server\nintroduces a code they do not know; use\n[`ErrorCode::parse`](crate::ErrorCode::parse) for typed access."
                },
                "details": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorDetails",
                      "description": "Structured context for the code, present when the failure carries\nmachine-usable identity (API spec, \"Standard error contract\"). Boxed\nso the rare detailed error does not widen every error-carrying result."
                    }
                  ]
                },
                "feature": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "For `not_supported` errors, the capability-document feature key the\nclient should reconcile against."
                },
                "message": {
                  "type": "string",
                  "description": "Human-readable error message."
                },
                "param": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "Identifies the invalid input. Body fields use JSON Pointer paths;\nquery and path parameters use their names; CLI errors use the flag or\nargument as written."
                },
                "request_id": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "Correlation id the server assigned to the failed request; the same\nvalue is sent as the `x-request-id` response header."
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The deployment's `auth_token`, sent as `Authorization: Bearer <token>`. A server configured without a token accepts every request; one configured with a token answers 401 `unauthorized` without it."
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}